Setup arc config and arc lint for development
Summary: I've also setup phpcs to be used by arc lint. Not sure what the codestyle of this project actually looks like. It looks a lot like the style that the PHPCS project follows itself, but by applying that I got almost a 1000 issues. So I disabled *a lot* of them. It went from 56 sniffs, to 16. We could have some discussion about the code style we want to use in this project?

Test Plan:
- That I was able to send it to Phabricator proves a part of this works. You can double check this by patching and running `arc which`, which also communicates with Phabricator.
- Make sure arc lint with phpcs works: Run `arc lint --everything`.
- You could have a look at the sniffs I've disabled. Remove/disable some of the excluded sniffs in phpcs.xml and run `arc lint --everything` again. See what new issues you get and whether you would like to keep them. We can also do this in a separate diff, leaving this only as a set up task (which obviously has my preference)

Reviewers: dennis, #uitzendbureau.nl

Reviewed By: dennis, #uitzendbureau.nl

Differential Revision: http://phabricator.hippo.farm/D2282