Updated PHPUnit to 7.5.20
Summary:
```
$ composer require --dev phpunit/phpunit:7.5.20 --update-with-dependencies
```

Test Plan:
Caution. This is a change to the {rHCS} project, so you'll need to open that first.

Patch this revision. Install the new phpunit version with `$ composer install`. Run the unit tests with `$ composer test` and `$ ./vendor/bin/phpunit`. Make sure nothing changes compared to the old version (same failures).

Reviewers: #review_round_1, armandas

Reviewed By: #review_round_1, armandas

Differential Revision: http://phabricator.hippo.farm/D6296